+++ lib/dart_core_extensions.dart
@@ -11,5 +11,5 @@ export 'src/file_size_label_extension.dart';
 export 'src/file_system_entity_core_extensions.dart';
 export 'src/int_size_label_extension.dart';
 export 'src/string_core_extensions.dart';
-export 'src/string_path_extensions.dart';
+export 'src/path_core_extensions.dart';
 export 'src/map_core_extensions.dart';