+++ lib/src/duration_eta_extensions.dart
@@ -13,4 +13,5 @@ extension DurationEtaExtensions on Duration {
   }
 
   
+  
 }