+++ lib/src/file_core_extensions.dart
@@ -1,15 +1,13 @@
 import 'dart:io';
 
-import 'package:dart_core_extensions/src/string_path_extensions.dart';
+import 'package:dart_core_extensions/src/path_core_extensions.dart';
 
 extension FileCoreExtensions on File {
   String getName({bool withExt = true}) {
     return path.getName(withExt: withExt);
   }
 
-  String get extName {
-    return path.extName;
-  }
+  String get extName => path.extName;
 
   String get parentPath => parent.path;
 
@@ -25,15 +23,7 @@ extension FileCoreExtensions on File {
     return statSync().size;
   }
 
-  Future<int> sizeAsync() async {
-    return (await stat()).size;
-  }
-
   DateTime get modified {
     return statSync().modified;
-  }
-
-  Future<DateTime> modifiedAsync() async {
-    return (await stat()).modified;
   }
 }