+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+
+- Added `MapCoreExtensions.putMapList`
+- Added `MapCoreExtensions.getMapList`
+- Added `PathCoreExtensions.join([path])`
+
 # 0.1.1
 
 - Added `MapCoreExtensions`